Change int type

+++ sorting/insertion-sort.cpp
@@ -6,7 +6,7 @@
 /* INSERTION SORT
 ============================================================================= */
 void insertionSort(std::vector<int>& arr) {
-    for (int i = 1; i < arr.size(); i++) {
+    for (size_t i = 1; i < arr.size(); i++) {
         int j = i - 1;
         while (j >= 0 && arr[j + 1] < arr[j]) {
             std::swap(arr[j], arr[j + 1]);
